Fixed Bug

+++ rcc_historia_persona.py
@@ -63,7 +63,7 @@ rcc.to_csv(os.path.join(path,rcc_file))
 d = {"cod_banco": [count_baks],'mto_saldo': ['mean','sum', 'std'], 'clasif': ['mean','min','max'], \
     'rango_mora': ['mean','min','max', rango_weak, rango_medium, rango_strong]} 
 rcc_banco = rcc.drop(['codmes'], axis=1).groupby(["id_persona", "producto"]).agg(d) \
-    .unstack(level=1, fill_value=0).reset_index().set_index("id_persona").sort_index().astype("int32")
+    .unstack(level=1, fill_value=0).reset_index().set_index("id_persona").sort_index().astype("float32")
 
 rcc_file = str('rcc_historia_persona.csv')
 rcc_banco.to_csv(os.path.join(path,rcc_file))
